feat(GenericContainer): extract environment variable handling to EnvSetting trait
Introduce the EnvSetting trait to encapsulate environment variable
management for containers. This refactoring removes duplicated code
from the GenericContainer class, promoting code reuse and improving
maintainability. The EnvSetting trait provides methods to add and
retrieve environment variables, ensuring consistent behavior across
containers. Additionally, new unit tests for EnvSetting are added to
validate its functionality, ensuring that environment variables are
correctly set and retrieved. This change enhances the modularity of
the codebase, allowing for easier future extensions and modifications.
